TIP-842: Move Attribute Group model
