+++ src/Pim/Bundle/CatalogBundle/Doctrine/ORM/Repository/AttributeGroupRepository.php
@@ -5,7 +5,7 @@ namespace Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository;
 use Doctrine\ORM\AbstractQuery;
 use Doctrine\ORM\EntityRepository;
 use Doctrine\ORM\QueryBuilder;
-use Pim\Bundle\CatalogBundle\Entity\AttributeGroup;
+use Akeneo\Pim\Structure\Component\Model\AttributeGroup;
 use Pim\Component\Catalog\Repository\AttributeGroupRepositoryInterface;
 
 /**
