+++ src/Pim/Bundle/CatalogBundle/Resources/config/doctrine/AttributeGroup.orm.yml
@@ -1,4 +1,4 @@
-Pim\Bundle\CatalogBundle\Entity\AttributeGroup:
+Akeneo\Pim\Structure\Component\Model\AttributeGroup:
     type: entity
     table: pim_catalog_attribute_group
     changeTrackingPolicy: DEFERRED_EXPLICIT
