+++ tests/back/Acceptance/AttributeGroup/InMemoryAttributeGroupRepository.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Akeneo\Test\Acceptance\AttributeGroup;
 
+use Akeneo\Pim\Structure\Component\Model\AttributeGroupInterface;
 use Akeneo\Tool\Component\StorageUtils\Saver\SaverInterface;
 use Akeneo\Test\Acceptance\Common\NotImplementedException;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
-use Pim\Component\Catalog\Model\AttributeGroupInterface;
 use Pim\Component\Catalog\Repository\AttributeGroupRepositoryInterface;
 
 /**
